Improved evaluation
CVEvaluator executes folds in parallel. Query- and MiningOptions are more explicit. Reduced disk usage when filtering types according to the number of projects they are used in (no need to actually count the usages of the types per project). Types are evaluated in alphabetical order. Results are stored in a map. Skipped types are counted differently.