Fixed a possible runtime exception during TypeCollector construction due to a method having two parameters that share the same name
The TypeCollectorVisitorContext now allows parameters to re-declare/overwrite previous entries in the same scope. Note that it would have also been possible to ignore the second parameter by catching the exception. However, this was not done in order to be consistent with the way normal variable declarations are handled.

Furthermore, the exception ScopedMap throws for duplicate keys has been specialized to facilitate controlled catching.